user delete and password change

+++ acmsummary/server/database.js
@@ -25,6 +25,7 @@ async function initDB(){
     await pool.query(`
         CREATE TABLE IF NOT EXISTS history (
             id SERIAL PRIMARY KEY,
+            userid INTEGER REFERENCES users(id) ON DELETE CASCADE,
             type VARCHAR(50) NOT NULL,
             title VARCHAR(255) UNIQUE NOT NULL,
             difficulty INTEGER,
@@ -50,31 +51,52 @@ const db = {
         return res.rows[0]
     },
 
-    async addSolvedQuestion(data){
+    async deleteUserById(userId){
         const res = await pool.query(`
-            INSERT INTO history (type, title, difficulty, solveTime) VALUES ($1, $2, $3, $4)
-        `, [data.type, data.title, data.difficulty, data.solveTime])
-        return res.rows[0]
-    },
-
-    async checkQuestionExistByTitle(title){
-        const res = await pool.query(`
-            SELECT * FROM history WHERE title = $1
-        `, [title])
-        return res.rows[0]
-    },
-
-    async getQuestionData(){
-        const res = await pool.query(`
-            SELECT * FROM history
-        `)
+            DELETE FROM users WHERE id = $1
+        `, [userId])
         return res
     },
 
-    async deleteQuestionDataById(id){
+    async getPasswordById(userId){
         const res = await pool.query(`
-            DELETE FROM history WHERE id = $1
-        `, [id])
+            SELECT password FROM users WHERE id = $1
+        `, [userId])
+        return res.rows[0]
+    },
+
+    async changePassword(userId, password){
+        const res = await pool.query(`
+            UPDATE users SET password = $1 WHERE id = $2
+        `, [password, userId])
+        return res
+    },
+
+    async addSolvedQuestion(data){
+        const res = await pool.query(`
+            INSERT INTO history (type, title, difficulty, solveTime, userid) VALUES ($1, $2, $3, $4, $5)
+        `, [data.type, data.title, data.difficulty, data.solveTime, data.userId])
+        return res.rows[0]
+    },
+
+    async checkQuestionExistById(id, userId){
+        const res = await pool.query(`
+            SELECT * FROM history WHERE title = $1 AND userid = $2
+        `, [id, userId])
+        return res.rows[0]
+    },
+
+    async getQuestionData(userId){
+        const res = await pool.query(`
+            SELECT * FROM history WHERE userid = $1
+        `, [userId])
+        return res
+    },
+
+    async deleteQuestionDataById(id, userId){
+        const res = await pool.query(`
+            DELETE FROM history WHERE id = $1 AND userid = $2
+        `, [id, userId])
         return res
     }
 }
